refactor(state): expose initialState to external code
